Fix bug
The bug are the cause that the program fails.
I fix some logic that avoid the problems of course, it have a lot. And I
now see that I don't have negative binaries.  The conversions between
bases now work fine and add the data to the stack.  The comands like
swap,negate will not work if the stack is empty. And more things like
that.